Fix: `window.asDynamic().navigator` to `window.navigator`
The `asDynamic()` call on `window` is no longer necessary to access the `navigator.vibrate()` API for Javascript and WasmJs targets.

Additionally, the `Navigator.vibrate()` external interface for WasmJs was simplified to accept `JsAny` as its parameter type, as the underlying JavaScript API can handle both single numbers and arrays of numbers.